lint stuff, move metrics code outside
